server/pyproject.toml
[project]
name = "dropbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "python-multipart",
    "bcrypt>=4.0.0",
    "pyjwt>=2.8.0",
    "pydantic-settings>=2.0.0"
]
